requirements.txt
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
faster-whisper>=1.0.0
soundfile>=0.12.1
pyttsx3>=2.90
ollama>=0.1.8
requests>=2.31.0
psutil>=5.9.0
pyperclip>=1.8.2
opencv-python>=4.8.0
pillow>=10.0.0
pyautogui>=0.9.54
pygame>=2.5.0
pystray>=0.19.5
pywebview>=6.1
websockets>=11.0
openwakeword>=0.6.0
deepface>=0.0.90
face_recognition>=1.3.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
